added board and gameboard
Board is basic with gameboard extending it with more data stored
